Pulled postFoodToMeal function into serviceIndex
