+++ lib/domIndex.js
@@ -1,4 +1,5 @@
-import { deleteMealFood, postFoodToMeal } from './index'
+import { deleteMealFood, } from './index'
+import { postFoodToMeal } from './serviceIndex'
 
 const mealCalories = {
   "breakfast": 400,
