+++ lib/index.js
@@ -1,7 +1,7 @@
 import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood, createFoodInMeal } from './domIndex'
 import { validateForm, getEachFood, removeFood, removeValidations, createInputField, removeInputField } from './domFoods'
 import { postFood, getAllFoodsValidation, createFood } from './serviceFoods'
-import { getAllFoodsDiary, countIncrement, getAllMeals } from './serviceIndex'
+import { getAllFoodsDiary, countIncrement, getAllMeals, postFoodToMeal } from './serviceIndex'
 
 export const handleResponse = (response) => {
   return response.json()
@@ -17,19 +17,6 @@ export const handleResponse = (response) => {
     }
     return json
   })
-}
-
-const postMealConfig = () => {
-  return {
-    method: 'POST',
-    headers: {'Content-Type': 'application/json'},
-  }
-}
-
-export const postFoodToMeal = (mealId, foodId) => {
-  fetch(`http://quantifiedself-backend.herokuapp.com/api/v1/meals/${mealId}/foods/${foodId}`, postMealConfig())
-    .then(() => getAllMeals())
-    .catch((error) => console.error({error}))
 }
 
 export const deleteMealFood = (foodId, mealId) => {
