+++ lib/serviceIndex.js
@@ -26,3 +26,16 @@ export const getAllMeals = () => {
     .then((meals) => appendMeals(meals))
     .catch((error) => console.error({error}))
 }
+
+export const postFoodToMeal = (mealId, foodId) => {
+  fetch(`http://quantifiedself-backend.herokuapp.com/api/v1/meals/${mealId}/foods/${foodId}`, postMealConfig())
+    .then(() => getAllMeals())
+    .catch((error) => console.error({error}))
+}
+
+const postMealConfig = () => {
+  return {
+    method: 'POST',
+    headers: {'Content-Type': 'application/json'},
+  }
+}
